requirements.txt
gradio
openai
transformers
numpy
Pillow
PyPDF2
arxiv
google-auth-oauthlib
google-api-python-client
google-auth-httplib2